feat: add user experiment when user clicked button

+++ app/[id]/page.tsx
@@ -4,28 +4,23 @@ import { useEffect, useState } from "react";
 import { IBook } from "@/types";
 import Header from "@/components/Header";
 import { useRouter } from "next/navigation";
-import {
-  deleteBookInDB,
-  getBookFromDB,
-  putBookInDB,
-} from "@/actions/actions";
+import { deleteBookInDB, getBookFromDB, putBookInDB } from "@/actions/actions";
 
 function BookPage({ params }: { params: { id: number } }) {
   const [book, setBook] = useState<IBook | null>(null);
   const [error, setError] = useState<string | null>(null);
+  const [loading, setLoading] = useState(false);
+  const [loadingDelete, setLoadingDelete] = useState(false);
   const router = useRouter();
 
   const handleChange = (
-    e: React.ChangeEvent<
-      HTMLInputElement | HTMLTextAreaElement
-    >
+    e: React.ChangeEvent<HTMLInputElement | HTMLTextAreaElement>
   ) => {
     const { name, value } = e.target;
     if (book) {
       const newBook: IBook = {
         ...book,
-        [name]:
-          name === "price" ? parseFloat(value) : value,
+        [name]: name === "price" ? parseFloat(value) : value,
       };
       setBook(newBook);
     }
@@ -33,6 +28,11 @@ function BookPage({ params }: { params: { id: number } }) {
 
   const handleEditSubmit = async (e: React.FormEvent) => {
     e.preventDefault();
+    const confirmed = window.confirm(
+      "Apakah Anda yakin ingin mengedit buku ini?"
+    );
+    if (!confirmed) return;
+    setLoading(true);
     try {
       if (book) {
         await putBookInDB(book);
@@ -42,11 +42,18 @@ function BookPage({ params }: { params: { id: number } }) {
     } catch (error) {
       console.log(error);
       setError("Failed to edit book");
+    } finally {
+      setLoading(false);
     }
   };
 
   const handleDeleteSubmit = async (e: React.FormEvent) => {
     e.preventDefault();
+    const confirmed = window.confirm(
+      "Apakah Anda yakin ingin menghapus buku ini?"
+    );
+    if (!confirmed) return;
+    setLoadingDelete(true);
     try {
       if (book) {
         await deleteBookInDB(params.id);
@@ -56,6 +63,8 @@ function BookPage({ params }: { params: { id: number } }) {
     } catch (error) {
       console.log(error);
       setError("Failed to delete book");
+    } finally {
+      setLoadingDelete(false);
     }
   };
 
@@ -85,16 +94,11 @@ function BookPage({ params }: { params: { id: number } }) {
             onSubmit={handleEditSubmit}
             className="my-5 w-full max-w-lg mx-auto"
           >
-            <h2 className="text-2xl font-bold mb-6">
-              Edit Book
-            </h2>
+            <h2 className="text-2xl font-bold mb-6">Edit Book</h2>
 
             {/* title */}
             <div className="mb-4">
-              <label
-                htmlFor="title"
-                className="block font-semibold"
-              >
+              <label htmlFor="title" className="block font-semibold">
                 Title
               </label>
               <input
@@ -111,10 +115,7 @@ function BookPage({ params }: { params: { id: number } }) {
 
             {/* author */}
             <div className="mb-4">
-              <label
-                htmlFor="author"
-                className="block font-semibold"
-              >
+              <label htmlFor="author" className="block font-semibold">
                 Author
               </label>
               <input
@@ -131,10 +132,7 @@ function BookPage({ params }: { params: { id: number } }) {
 
             {/* price */}
             <div className="mb-4">
-              <label
-                htmlFor="price"
-                className="block font-semibold"
-              >
+              <label htmlFor="price" className="block font-semibold">
                 Price
               </label>
               <input
@@ -152,10 +150,7 @@ function BookPage({ params }: { params: { id: number } }) {
 
             {/* description */}
             <div className="mb-4">
-              <label
-                htmlFor="description"
-                className="block font-semibold"
-              >
+              <label htmlFor="description" className="block font-semibold">
                 Description
               </label>
               <textarea
@@ -172,8 +167,9 @@ function BookPage({ params }: { params: { id: number } }) {
             <button
               type="submit"
               className="w-full py-3 bg-btn-color text-white rounded hover:bg-text-hover transition duration-200"
+              disabled={loading}
             >
-              Edit Book
+              {loading ? "Loading..." : "Edit Book"}
             </button>
           </form>
           <form
@@ -183,8 +179,9 @@ function BookPage({ params }: { params: { id: number } }) {
             <button
               type="submit"
               className="w-full py-3 bg-red-700 text-white rounded hover:bg-red-600 transition duration-200"
+              disabled={loadingDelete}
             >
-              Delete Book
+              {loadingDelete ? "Loading..." : "Delete Book"}
             </button>
           </form>
         </div>
